Changes to classes CRUD in order to deal with both Exercicio and Atendente in the database

+++ demo/src/main/java/br/cefet/model/Exercicio.java
@@ -37,7 +37,7 @@ public class Exercicio {
         return carga;
     }
 
-    public float getTempoDescanso(){
+    public int getTempoDescanso(){
         return tempoDescanso;
     }
